refactor: move snapshot output formatting to snapshots/output.py
